g_RunningSpellcardInfo is part of EclManager
